requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
redis==5.0.1
pypdf==4.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
pydantic==2.6.1 